deployment

+++ controllers/postController.js
@@ -15,83 +15,89 @@ const authenticateAdmin = require('../config/jwtMiddleware');
 
 module.exports.PostConroller = async function(req,res){
 
-    Post.uploadPicture(req,res,async function(err){
+    try {
+       Post.uploadPicture(req,res,async function(err){
 
-        // console.log(req.user);
-        if(err){
-            console.error(err);
-        }
-        const user_type = req.query.type;
-
-        var hotelName ;
-        var menuModel ;
-
-        if(user_type=="User"){
-            hotelName = req.body.hotel;
-            menuModel = req.body.menuItem;
-        }
-        console.log("----------------------------------");
-        let presentPost = await Post.create({
-            user:req.user._id,
-            content:req.body.content,
-            picturePath:Post.picPath + "/"+req.file.filename,
-            upVotesCount:req.body.upVotesCount,
-            UserOrHotel:req.body.userType,
-            hotelName : hotelName,
-            menuModel : menuModel
-        });
-
-        
-
-        let post_User;
-
-        if(user_type=="User"){
-            
-            post_User = await User.findById(req.user._id);
-            
-            if(req.body.hotel){
-
-                const hotel = await Hotel.findById(req.body.hotel);
-
-                hotel.reviewPosts.push(presentPost.id);
-
-                await hotel.save();
-
+            // console.log(req.user);
+            if(err){
+                console.error(err);
+    
             }
-
-            post_User.posts.push(presentPost.id);             
-
-        }
-        else{
-
-            post_User = await Hotel.findById(req.user._id);
-
-            post_User.posts.push(presentPost.id);
+            const user_type = req.query.type;
+    
+            var hotelName ;
+            var menuModel ;
+    
+            if(user_type=="User"){
+                hotelName = req.body.hotel;
+                menuModel = req.body.menuItem;
+            }
+            console.log("----------------------------------");
+            let presentPost = await Post.create({
+                user:req.user._id,
+                content:req.body.content,
+                picturePath:Post.picPath + "/"+req.file.filename,
+                upVotesCount:req.body.upVotesCount,
+                UserOrHotel:req.body.userType,
+                hotelName : hotelName,
+                menuModel : menuModel
+            });
+    
             
-        }
+    
+            let post_User;
+    
+            if(user_type=="User"){
+                
+                post_User = await User.findById(req.user._id);
+                
+                if(req.body.hotel){
+    
+                    const hotel = await Hotel.findById(req.body.hotel);
+    
+                    hotel.reviewPosts.push(presentPost.id);
+    
+                    await hotel.save();
+    
+                }
+    
+                post_User.posts.push(presentPost.id);             
+    
+            }
+            else{
+    
+                post_User = await Hotel.findById(req.user._id);
+    
+                post_User.posts.push(presentPost.id);
+                
+            }
+            
+            let upvote = await upVote.create({
+                votable:presentPost.id,
+                postORcomment:"Post",
+                user : req.user.id,
+                upVoted:false,
+                UserOrHotel:req.body.userType
+            });
+    
+            presentPost.upVotes.push(upvote.id);
+    
+            await presentPost.save();
+    
+            await post_User.save();
+    
+            await req.flash('message', [
+                { type: 'flash-success', text: 'Post Created' },
+              ]);
         
-        let upvote = await upVote.create({
-            votable:presentPost.id,
-            postORcomment:"Post",
-            user : req.user.id,
-            upVoted:false,
-            UserOrHotel:req.body.userType
+            return res.redirect('/');
         });
+    } catch (error) {
+        console.log(error);
+    }
+    
 
-        presentPost.upVotes.push(upvote.id);
-
-        await presentPost.save();
-
-        await post_User.save();
-
- 
-    });
-
-    await req.flash('message', [
-        { type: 'flash-success', text: 'Post Created' },
-      ]);
-
-    return res.redirect('/');
+    
 }
 
 module.exports.deletePost = async function(req,res,next){
@@ -119,42 +125,50 @@ module.exports.deletePost = async function(req,res,next){
         select:'id'
     });
     
+    if(post){
+        // console.log("in this", req.user.id,post.user.id,req.user.isAdmin);
+        if(req.user.id == post.user.id || req.user.isAdmin){
     
-    // console.log("in this", req.user.id,post.user.id,req.user.isAdmin);
-    if(req.user.id == post.user.id || req.user.isAdmin){
-    
-        let commentData = post.comments;
-        
-        if (commentData.length > 0){
-            for (const comment of commentData){
-                await upVote.deleteOne({votable:comment._id});
-                await Comment.deleteOne({id:comment._id});
+            let commentData = post.comments;
+            
+            if (commentData.length > 0){
+                for (const comment of commentData){
+                    await upVote.deleteOne({votable:comment._id});
+                    await Comment.deleteOne({id:comment._id});
+                }
             }
+        
+            await PostUser.posts.pull(req.params.pID);
+        
+            await upVote.deleteMany({votable:req.params.pID});
+        
+            // unlink the image
+            const oldProfilePath = path.join(__dirname,'../',post.picturePath);
+            fs.unlinkSync(oldProfilePath);
+    
+            await post.remove();
+        
+        
+            await PostUser.save();
+            await req.flash('message', [
+                { type: 'flash-warning', text: 'Post Deleted' },
+              ]);
+            return res.redirect('back');
         }
-    
-        await PostUser.posts.pull(req.params.pID);
-    
-        await upVote.deleteMany({votable:req.params.pID});
-    
-        // unlink the image
-        const oldProfilePath = path.join(__dirname,'../',post.picturePath);
-        fs.unlinkSync(oldProfilePath);
-
-        await post.remove();
-    
-    
-        await PostUser.save();
-        await req.flash('message', [
-            { type: 'flash-warning', text: 'Post Deleted' },
-          ]);
-        return res.redirect('back');
+        else{
+            await req.flash('message', [
+                { type: 'flash-warning', text: "Can't delete" },
+              ]);
+            return res.redirect('back');
+        }
     }
     else{
         await req.flash('message', [
-            { type: 'flash-warning', text: "Can't delete" },
+            { type: 'flash-warning', text: "Post Unavailable" },
           ]);
         return res.redirect('back');
     }
+    
 
 }
 
@@ -162,6 +176,12 @@ module.exports.download = async function(req,res){
     const postId = req.params.id;
 
     const currentPost =  await Post.findById(postId);
+    if(currentPost==null){
+        await req.flash('message', [
+            { type: 'flash-warning', text: "Post Unavailable !" },
+          ]);
+          return res.redirect('back');
+    }
 
     const imgPath = path.join(__dirname,'../',currentPost.picturePath);
     const outputPath = path.join(__dirname,'../uploads/buffer/');
@@ -188,6 +208,9 @@ module.exports.download = async function(req,res){
     const downloadFilePath =path.join(__dirname,'../uploads/buffer/',uniqueFilename);
 
     // console.log(downloadFilePath);
+    // await req.flash('message', [
+    //     { type: 'flash-success', text: 'Download Success' },
+    //   ]);
     await res.download(downloadFilePath, (err) => {
         if (err) {
           console.error('Error downloading image:', err);
